+++ src/index.ts
@@ -4,7 +4,8 @@ import { container } from "./infrastructure/ioc/Container";
 import { MeteorologicController } from "./controllers/MeteorologicController";
 import { Logger } from "winston";
 import { TYPES } from "./infrastructure/ioc/Types";
-
+import * as pack from "../package.json";
+const prefix = "[Meteorologic Collector]";
 useContainer(container);
 
 const logger = container.get<Logger>(TYPES.Logger);
@@ -17,7 +18,8 @@ const app = createExpressServer({
 });
 
 app.listen(process.env.PORT || 8080, () => {
-  logger.info(
-    `Meteorologic Collector listen on port ${process.env.PORT || 8080}`,
-  );
+  logger
+    .info(`${prefix} started with version: v${pack.version}`)
+    .info(`${prefix} log level is set to ${process.env.LOG_LEVEL ?? "debug"}`)
+    .info(`${prefix} listen on port ${process.env.PORT ?? 8080}`);
 });